api/pyproject.toml
[tool.poetry]
name = "studx-api"
version = "0.1.0"
description = ""
authors = ["Hans Tognon <learningway403@gmail.com>"]
readme = "README.md"
packages = [{ include = "studx_api" }]

[tool.poetry.dependencies]
python = "^3.10"
django-celery-beat = "^2.4.0"
djangorestframework = "^3.14.0"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.5"
django-rest-knox = "^4.2.0"
pyjwt = "^2.6.0"
drf-spectacular = "^0.25.1"
djangorestframework-camel-case = "^1.3.0"
pillow = "^9.4.0"
django-notifications-hq = "^1.7.0"
redis = "^4.4.2"
drf-nested-routers = "^0.93.4"


[tool.poetry.group.dev.dependencies]
django-extensions = "^3.2.1"
django-stubs = "^1.13.1"
djangorestframework-stubs = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
